requirements.txt
Flask==1.1.1
flask-cors==3.0.10
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
Werkzeug==0.15.5
numpy==1.25.2
scipy==1.11.4
scikit-learn==1.2.2
matplotlib==3.7.1
sklearn-pandas==2.2.0

joblib==1.4.2
xgboost==2.0.3
pyod==1.1.3
gunicorn==20.1.0
json5==0.9.5
regex==2023.12.25
